requirements.txt
aiohttp>=3.9,<4
python-dotenv>=1.0,<2
discord.py>=2.3,<3
certifi>=2024.7.4
